pyproject.toml
[tool.poetry]
name = "gul"
version = "0.0.1"
description = "Grand Unified Lookup"
authors = ["Utah State University"]
license = "MIT"

[tool.isort]
profile = "black"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
python-dotenv = "^0.20.0"
databases = {extras = ["asyncpg"], version = "^0.6.0"}
sqlmodel = "^0.0.6"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
safety = "^1.10.3"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
